Extract verifier_repair_pass_flow.rs from turn.rs (parent #680)
Move the per-attempt verifier-repair pass flow (the repair-plan-driven
targeted-edit loop) out of `turn.rs` into a focused sibling module so
`turn.rs` keeps shrinking toward responsibility-focused units (parent
meta-issue #680).

Three production helpers + two private helpers extracted as free
functions taking `&mut Agent` / `&Agent` (matching `actor_loop_flow` /
`reply_retry` / earlier vertical-slice precedent):

- `prepare_verifier_repair_pass` (admission gate + accepted-plan build
  + behavior-projection emit + prompt-message render)
- `handle_verifier_repair_pass_attempt` (per-attempt dispatcher →
  `VerifierRepairAttemptProgress`)
- `verifier_repair_pass_wall_clock_timeout_error` (timeout error +
  bounded log emit)
- `handle_verifier_repair_pass_reply` (private): patch-proposal parse
  + shadow validation + intent admission + legacy comparison emit +
  apply via `verifier_orchestration::apply_verifier_repair_pass_edit`
- `log_verifier_repair_pass_timeout` (private): bounded
  `agent.verifier_repair_pass.timeout` event

Behavior unchanged: same admission gate ordering, same accepted-plan
build + `validate_accepted_repair_plan_authorizes_target` guard, same
shadow→legacy validation comparison emit, same per-attempt error
classification + `VerifierRepairAttemptProgress` mapping, same
`build_verifier_repair_pass_ledger_outcome` lineage for invalid /
weakening / rejection signals.

`pub(super)` limited / no facade re-export (DR3-001). Callers updated
in `verifier_orchestration.rs` (3 sites) to the free-fn form.

LOC delta:
- turn.rs: 2,200 → 1,923 (-277)
- new module: +300

3,114 lib tests pass; `cargo fmt` / `cargo clippy --lib --all-targets
-- -D warnings` clean.
